backend/requirements.txt
aniso8601==10.0.0
annotated-types==0.7.0
attrs==25.1.0
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
exceptiongroup==1.2.2
gevent==21.12.0 
gunicorn==21.2.0
Flask==3.1.0
flask-restx==1.3.0
flask-cors==5.0.1
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.2
pytest==8.3.5
python-dotenv==1.0.1
pytz==2025.1
referencing==0.36.2
rpds-py==0.23.1
tomli==2.2.1
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
bcrypt==4.3.0
pyjwt==2.10.1
